+++ katherine-fundraising-contract/src/internal.rs
@@ -9,14 +9,6 @@ use crate::interface::*;
 /*************/
 
 impl KatherineFundraising {
-    pub(crate) fn assert_min_deposit_amount(&self, amount: Balance) {
-        assert!(
-            amount >= self.min_deposit_amount,
-            "minimum deposit amount is {}",
-            self.min_deposit_amount
-        );
-    }
-
     pub(crate) fn assert_unique_slug(&self, slug: &String) {
         assert!(
             self.kickstarter_id_by_slug.get(slug).is_none(),
@@ -452,60 +444,6 @@ impl KatherineFundraising {
         self.kickstarters
             .get(kickstarter_id as u64)
             .expect("Unknown KickstarterId")
-    }
-
-    /// Process a stNEAR deposit to Katherine Contract.
-    pub(crate) fn internal_supporter_deposit(
-        &mut self,
-        supporter_id: &AccountId,
-        amount: &Balance,
-        kickstarter: &mut Kickstarter,
-    ) {
-        // Update Kickstarter
-        kickstarter.assert_within_funding_period();
-        kickstarter.assert_enough_reward_tokens();
-
-        let new_total_deposited = kickstarter.total_deposited + amount;
-        assert!(
-            new_total_deposited <= kickstarter.deposits_hard_cap,
-            "The deposits hard cap cannot be exceeded!"
-        );
-        kickstarter.total_deposited = new_total_deposited;
-        kickstarter.update_supporter_deposits(&supporter_id, amount);
-        self.kickstarters
-            .replace(kickstarter.id as u64, &kickstarter);
-
-        // Update Supporter.
-        let mut supporter = self.internal_get_supporter(&supporter_id);
-        supporter.total_in_deposits += amount;
-        supporter.supported_projects.insert(&kickstarter.id);
-        self.supporters.insert(&supporter_id, &supporter);
-    }
-
-    /// Process a reward token deposit to Katherine Contract.
-    pub(crate) fn internal_kickstarter_deposit(
-        &mut self,
-        amount: &Balance,
-        kickstarter: &mut Kickstarter,
-    ) {
-        assert_eq!(
-            &env::predecessor_account_id(),
-            &kickstarter.token_contract_address,
-            "Deposited tokens do not correspond to the Kickstarter contract."
-        );
-        assert!(
-            get_current_epoch_millis() < kickstarter.close_timestamp,
-            "Kickstarter Tokens should be provided before the funding period ends."
-        );
-        let max_tokens_to_release = self.calculate_max_tokens_to_release(&kickstarter);
-        let min_tokens_to_allow_support = max_tokens_to_release
-            + self.calculate_katherine_fee(max_tokens_to_release);
-        kickstarter.available_reward_tokens += amount;
-        kickstarter.enough_reward_tokens = {
-            kickstarter.available_reward_tokens >= min_tokens_to_allow_support
-        };
-        self.kickstarters
-            .replace(kickstarter.id as u64, &kickstarter);
     }
 
     pub(crate) fn activate_successful_kickstarter(
